+++ dogs_detection/android/src/com/beginvision/dogdetector/NativeAgent.java
@@ -9,7 +9,7 @@ import android.util.Log;
 public class NativeAgent{
     public static native int init();
     public static native int updatePicture(byte[]frame, int wid, int hei);
-    public static native int updatePictureForResult(byte[]frame, Object obj, int wid, int hei);
+    public static native double updatePictureForResult(byte[]frame, Object obj, int wid, int hei);
     static {
         System.loadLibrary("beginvision");
     }
